+++ FE/src/page/interviewPage/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useRef, useEffect, useCallback } from 'react';
+import React, { useState } from 'react';
 
 import InterviewPageLayout from '@components/interviewPage/InterviewPageLayout';
 import InterviewHeader from '@/components/interviewPage/InterviewHeader/InterviewHeader';
@@ -6,88 +6,32 @@ import InterviewMain from '@/components/interviewPage/InterviewMain/InterviewMai
 import InterviewFooter from '@/components/interviewPage/InterviewFooter/InterviewFooter';
 import InterviewIntroModal from '@components/interviewPage/InterviewModal/InterviewIntroModal';
 import InterviewTimeOverModal from '@components/interviewPage/InterviewModal/InterviewTimeOverModal';
-import useInterviewFlow from '@hooks/pages/Interview/useInterviewFlow';
-import useIsAllSuccess from '@/hooks/pages/Interview/useIsAllSuccess';
+
 import { PATH } from '@constants/path';
 import { Navigate } from 'react-router-dom';
-import { useRecoilValue } from 'recoil';
-import { recordSetting } from '@/atoms/interviewSetting';
-import useMedia from '@/hooks/useMedia';
-import { localDownload, startRecording, stopRecording } from '@/utils/record';
-import { useUploadToIDrive } from '@/hooks/pages/Interview/useUploadToIdrive';
-import useTimeTracker from '@/hooks/pages/Interview/useTimeTracker';
 
+import useInterview from '@/hooks/pages/Interview/useInterview';
 const InterviewPage: React.FC = () => {
-  const { startTimer, stopTimer, calculateDuration, isTimeOver } =
-    useTimeTracker();
-  const isAllSuccess = useIsAllSuccess();
-  const { method } = useRecoilValue(recordSetting);
-  const uploadToDrive = useUploadToIDrive();
-  const { currentQuestion, getNextQuestion, isLastQuestion } =
-    useInterviewFlow();
-
   const {
-    media,
-    videoRef: mirrorVideoRef,
+    isAllSuccess,
     connectStatus,
-    selectedMimeType,
-  } = useMedia();
+    isRecording,
+    videoRef,
+    isScriptInView,
+    setIsScriptInView,
+    recordedBlobs,
+    currentQuestion,
+    isLastQuestion,
+    getNextQuestion,
+    handleStartRecording,
+    handleStopRecording,
+    handleDownload,
+    timeOverModalIsOpen,
+    setTimeOverModalIsOpen,
+  } = useInterview();
 
-  const [isRecording, setIsRecording] = useState(false);
-  const [isScriptInView, setIsScriptInView] = useState(true);
-  const [recordedBlobs, setRecordedBlobs] = useState<Blob[]>([]);
   const [interviewIntroModalIsOpen, setInterviewIntroModalIsOpen] =
     useState<boolean>(true);
-  const [interviewTimeOverModalIsOpen, setInterviewTimeOverModalIsOpen] =
-    useState<boolean>(false);
-
-  const mediaRecorderRef = useRef<MediaRecorder | null>(null);
-
-  const handleStartRecording = useCallback(() => {
-    startRecording({
-      media,
-      selectedMimeType,
-      mediaRecorderRef,
-      setRecordedBlobs,
-    });
-    setIsRecording(true);
-    startTimer();
-  }, [media, selectedMimeType, mediaRecorderRef, setRecordedBlobs, startTimer]);
-
-  const handleStopRecording = useCallback(() => {
-    stopRecording(mediaRecorderRef);
-    setIsRecording(false);
-    stopTimer();
-  }, [mediaRecorderRef, stopTimer]);
-
-  const handleDownload = useCallback(() => {
-    const blob = new Blob(recordedBlobs, { type: selectedMimeType });
-    const recordingTime = calculateDuration();
-    console.log(recordingTime);
-    switch (method) {
-      case 'idrive':
-        void uploadToDrive({ blob, currentQuestion });
-        break;
-      case 'local':
-        localDownload(blob, currentQuestion);
-        break;
-    }
-    setRecordedBlobs([]);
-  }, [
-    recordedBlobs,
-    selectedMimeType,
-    method,
-    currentQuestion,
-    uploadToDrive,
-    calculateDuration,
-  ]);
-
-  useEffect(() => {
-    if (isTimeOver) {
-      setInterviewTimeOverModalIsOpen(true);
-      handleStopRecording();
-    }
-  }, [handleStopRecording, isTimeOver]);
 
   if (!isAllSuccess || connectStatus === 'fail') {
     return <Navigate to={PATH.ROOT} />;
@@ -96,7 +40,7 @@ const InterviewPage: React.FC = () => {
       <InterviewPageLayout>
         <InterviewHeader isRecording={isRecording} intervieweeName="면접자" />
         <InterviewMain
-          mirrorVideoRef={mirrorVideoRef}
+          mirrorVideoRef={videoRef}
           isScriptInView={isScriptInView}
           question={currentQuestion.questionContent}
           answer={currentQuestion.answerContent}
@@ -116,8 +60,8 @@ const InterviewPage: React.FC = () => {
           closeModal={() => setInterviewIntroModalIsOpen((prev) => !prev)}
         />
         <InterviewTimeOverModal
-          isOpen={interviewTimeOverModalIsOpen}
-          closeModal={() => setInterviewTimeOverModalIsOpen(false)}
+          isOpen={timeOverModalIsOpen}
+          closeModal={() => setTimeOverModalIsOpen(false)}
         />
       </InterviewPageLayout>
     );
